Версия: 1.3.0: 1) Теперь более удобный конструктор модальных окон на ES6 Class 2) Добавлена модалка feedback
