+++ src/js/index.js
@@ -1,6 +1,6 @@
 import "../scss/styles.scss";
 import "./script.js"
-
-
+import "./swiper.js"
+import 'loading-attribute-polyfill';
