+++ src/js/script.js
@@ -1,85 +1,52 @@
-import Swiper, { Pagination } from "swiper";
-Swiper.use(Pagination);
+const wrapper = document.querySelector(".wrapper");
+const buttonEscape = 27;
 
-const menuController = (() => {
-    const wrapper = document.querySelector(".wrapper");
-    const menu = wrapper.querySelector(".menu");
-    const openButton = wrapper.querySelector("#openMenu");
-    const closeButton = wrapper.querySelector("#closeMenu *");
-
-    function menuOpenHandler() {
-        openButton.addEventListener(
-            "click",
-            (e) => {
-                e.preventDefault();
-                menu.classList.add("menu--active");
-                wrapper.classList.add("wrapper--disabled");
-                awaitClose();
-            },
-            { once: true }
-        );
+class ModalConstructor {
+    constructor(element, activeElement, openButton, closeButton) {
+        this.element = wrapper.querySelector(element);
+        this.activeElement = activeElement;
+        this.openButton = wrapper.querySelector(openButton);
+        this.closeButton = wrapper.querySelector(closeButton);
     }
 
-    function menuCloseHandler(e) {
-        e.preventDefault();
-        if (e.keyCode === 27 || e.target === wrapper || e.target === closeButton) {
-            window.removeEventListener("keydown", menuCloseHandler);
-            window.addEventListener("click", menuCloseHandler);
-            menu.classList.remove("menu--active");
-            wrapper.classList.remove("wrapper--disabled");
+    activate() {
+        const awaitModalOpen = () => {
+            this.openButton.addEventListener(
+                "click",
+                (e) => {
+                    e.preventDefault();
+                    this.element.classList.add(this.activeElement);
+                    wrapper.classList.add("wrapper--disabled");
+                    awaitClose();
+                },
+                { once: true }
+            );
+        };
 
-            menuOpenHandler();
-        }
-    }
+        const awaitModalClose = (e) => {
+            e.preventDefault();
+            if (e.keyCode === buttonEscape || e.target === wrapper || e.target === this.closeButton) {
+                window.removeEventListener("keydown", awaitModalClose);
+                window.addEventListener("click", awaitModalClose);
+                this.element.classList.remove(this.activeElement);
+                wrapper.classList.remove("wrapper--disabled");
 
-    function awaitClose() {
-        window.addEventListener("keydown", menuCloseHandler);
-        window.addEventListener("click", menuCloseHandler);
-    }
-
-    menuOpenHandler();
-})();
-
-const initSwiper = (() => {
-    const breakpoint = window.matchMedia("(min-width:768px)");
-
-    let mySwiper;
-
-    const breakpointChecker = () => {
-        if (breakpoint.matches === true) {
-            if (mySwiper !== undefined) {
-                mySwiper.destroy(true, true);
-            } else {
-                return;
+                awaitModalOpen();
             }
-        } else if (breakpoint.matches === false) {
-            return enableSwiper();
-        }
-    };
+        };
 
-    const enableSwiper = function () {
-        mySwiper = new Swiper(".swiper", {
-            slidesPerView: "auto",
-            loop: false,
-            a11y: true,
-            spaceBetween: 16,
-            slideClass: "swiper__slide",
-            slideActiveClass: "swiper__slide--active",
-            wrapperClass: "swiper__wrapper",
-            containerModifierClass: "swiper--",
-            slideNextClass: "swiper__slide--next",
-            slidePrevClass: "swiper__slide--prev",
-            pagination: {
-                el: ".swiper__pagination",
-                clickable: false,
-                modifierClass: "swiper__pagination--",
-                bulletClass: "swiper__bullet",
-                bulletActiveClass: "swiper__bullet--active",
-            },
-        });
-    };
+        const awaitClose = () => {
+            window.addEventListener("keydown", awaitModalClose);
+            window.addEventListener("click", awaitModalClose);
+        };
+        awaitModalOpen();
+    }
+}
 
-    breakpoint.addListener(breakpointChecker);
+const menu = new ModalConstructor(".menu", "menu--active", "#openMenu", "#closeMenu *");
+menu.activate();
 
-    breakpointChecker();
-})();
+const feedback = new ModalConstructor(".feedback", "feedback--active", "#openFeedback", "#closeForm *");
+feedback.activate();
+
+
