+++ src/scss/Blocks/Blocks__Bundle.scss
@@ -8,3 +8,4 @@
 @import "./menu";
 @import "./brends";
 @import "./swiper";
+@import "./feedback";
