+++ src/scss/Blocks/menu.scss
@@ -18,6 +18,7 @@
 
         &--active {
             left: 0;
+            height: 100%;
         }
     }
